requirements.txt
Pillow
pycairo
